+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+### 20200913
+- Added a few extra missing figures
+- Added grating stimuli: `GratingDirection`, `GratingSpatial`, `GratingTemporal`
+- Cleaned up `BarCentering`
+
 ### 20200912 
 - Moved over relevant code from [patterson-package][old_protocols]